moving table sorting vue components to their own folder
